Gridrow charts

+++ FinalCapstone/Controllers/CallSessionController.cs
@@ -1,6 +1,7 @@
 ﻿using System;
 using System.Collections.Generic;
 using System.Linq;
+using System.Security.Claims;
 using System.Threading.Tasks;
 using FinalCapstone.Data;
 using FinalCapstone.Models;
@@ -16,66 +17,100 @@ namespace FinalCapstone.Controllers
     public class CallSessionController : ControllerBase
     {
         private readonly CallSessionRepository _callSessionRepo;
+        private readonly UserProfileRepository _userProfileRepo;
         public CallSessionController(ApplicationDBContext context)
         {
             _callSessionRepo = new CallSessionRepository(context);
+            _userProfileRepo = new UserProfileRepository(context);
         }
 
-        [HttpGet("{id}")]
-        public IActionResult Get(int id, int days)
+        [HttpGet]
+        public IActionResult Get(int days)
         {
-
+            var currentUser = GetCurrentUserProfile();
             var startdate = DateTime.Now - TimeSpan.FromDays(days);
-            var callSessions = _callSessionRepo.GetByUserId(id, startdate);
-            if (callSessions == null)
-            {
-                return NotFound();
-            }
+            var callSessions = _callSessionRepo.GetByUserId(currentUser.Id, startdate);
             return Ok(callSessions);
+ 
         }
 
         [HttpPost]
         public IActionResult CallSession(CallSession callsession)
         {
-            _callSessionRepo.Add(callsession);
-            return CreatedAtAction("Get", new { id = callsession.Id }, callsession);
+            var currentUser = GetCurrentUserProfile();
+            callsession.UserProfileId = currentUser.Id;
+
+                _callSessionRepo.Add(callsession);
+                return CreatedAtAction("Get", new { id = callsession.Id }, callsession);
+
         }
 
-        [HttpPut("{id}")]
-        public IActionResult Put(int id, CallSession callsession)
+        [HttpPut]
+        public IActionResult Put(CallSession callsession)
         {
-            if (id != callsession.Id)
+            var currentUser = GetCurrentUserProfile();
+            if (callsession.UserProfileId == currentUser.Id)
             {
-                return BadRequest();
+                _callSessionRepo.Update(callsession);
+                return NoContent();
+            }
+            else
+            {
+                return Unauthorized();
             }
 
-            _callSessionRepo.Update(callsession);
-            return NoContent();
         }
 
         [HttpDelete("{id}")]
         public IActionResult Delete(int id)
         {
-            _callSessionRepo.Delete(id);
-            return NoContent();
+            var callSession = _callSessionRepo.GetById(id);
+            var currentUser = GetCurrentUserProfile();
+            if(callSession.UserProfileId == currentUser.Id)
+                {
+                    _callSessionRepo.Delete(id);
+                    return NoContent();
+                }
+            else
+            {
+                return Unauthorized();
+            }
         }
 
-        [HttpGet("{id}/contactratio")]
-        public IActionResult GetContactRatio(int id, int days)
+        [HttpGet("contactratio")]
+        public IActionResult GetContactRatio(int days)
         {
+            var currentUser = GetCurrentUserProfile();
             var startdate = DateTime.Now - TimeSpan.FromDays(days);
-            var contacts = _callSessionRepo.GetContactsTotal(id, startdate);
-            var appointmentsBooked = _callSessionRepo.GetAppointmentBookedTotal(id, startdate);
-
-
-
+            var contacts = _callSessionRepo.GetContactsTotal(currentUser.Id, startdate);
+            var appointmentsBooked = _callSessionRepo.GetAppointmentBookedTotal(currentUser.Id, startdate);
             var appointmentRatioView = new ContactRatioViewModel()
             {
                 Contacts = contacts,
                 AppointmentsBooked = appointmentsBooked
             };
+
+
             return Ok(appointmentRatioView);
 
+        }
+
+        [HttpGet("callcomparison")]
+        public IActionResult GetCallComparison(int days)
+        {
+            var startdate = DateTime.Now - TimeSpan.FromDays(days);
+            var currentUser = GetCurrentUserProfile();
+
+            var callComparison = _callSessionRepo.GetCallComparisonByUserId(currentUser.Id, startdate);
+
+
+            return Ok(callComparison);
+
+        }
+        private UserProfile GetCurrentUserProfile()
+        {
+            var firebaseUserId = User.FindFirst(ClaimTypes.NameIdentifier).Value;
+            return _userProfileRepo.GetByFirebaseId(firebaseUserId);
         }
 
     }
